+++ app/src/main/java/ru/yandex/yamblz/ui/recycle_view_beatifiers/OnItemDragListener.java
@@ -1,4 +1,4 @@
-package ru.yandex.yamblz.ui.adapters;
+package ru.yandex.yamblz.ui.recycle_view_beatifiers;
 
 /**
  * Created by Volha on 28.07.2016.
