+++ app/src/main/java/ru/yandex/yamblz/ui/recycle_view_beatifiers/ScrollAnimationListener.java
@@ -1,11 +1,7 @@
-package ru.yandex.yamblz.ui.adapters;
+package ru.yandex.yamblz.ui.recycle_view_beatifiers;
 
-import android.content.Context;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.AttributeSet;
-import android.util.Log;
-import android.view.View;
 
 /**
  * Created by Volha on 31.07.2016.
